refactor(extract): move resolve options up in the call hierarchy
